product type controller done

+++ BangazonAPI/Models/Product.cs
@@ -8,19 +8,15 @@ namespace BangazonAPI.Models
 {
     public class Product
     {
+
         public int Id { get; set; }
         public string Title { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string Description { get; set; }
         //CustomerId is the seller
         public int CustomerId { get; set; }
         public int ProductTypeId { get; set; }
         public DateTime DateAdded { get; set; }
-
-
-
-
-
 
     }
 }
